requirements.txt
flask
scikit-learn
joblib
pandas
openai
xhtml2pdf
numpy
matplotlib
python-dotenv
requests
